Splitted Image\Value into Base,Value and Input
- Base is abstract, and holds user-defined attributes (alternative text & fileName)
  + fileSize
- Value is built only by the API, and contains id, imageId and uri
- InputValue is used for inputing a new image, and adds an imageUri property.
  Also sets fileSize to the size of the provided file.
